Fix JS export and move utils into ExprUtils
