+++ src/Language/Mimsa/Interpreter/HighestVar.hs
@@ -1,6 +1,7 @@
 module Language.Mimsa.Interpreter.HighestVar (highestVar) where
 
 import Data.Semigroup (Max (..))
+import Language.Mimsa.ExprUtils
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Identifiers
 
@@ -9,6 +10,6 @@ import Language.Mimsa.Types.Identifiers
 highestVar :: Expr Variable ann -> Int
 highestVar expr' = max 0 $ getMax $ withMonoid getHighest expr'
   where
-    getHighest (MyVar _ (NumberedVar i)) = Max i
-    getHighest (MyDefineInfix _ _ (NumberedVar i) _) = Max i
-    getHighest _ = mempty
+    getHighest (MyVar _ (NumberedVar i)) = (False, Max i)
+    getHighest (MyDefineInfix _ _ (NumberedVar i) _) = (True, Max i)
+    getHighest _ = (True, mempty)
