+++ src/Language/Mimsa/Interpreter/SwapName.hs
@@ -5,6 +5,7 @@ module Language.Mimsa.Interpreter.SwapName
   )
 where
 
+import Language.Mimsa.ExprUtils
 import Language.Mimsa.Interpreter.Types
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Identifiers
