+++ test/Test/Backend/BackendJS.hs
@@ -79,7 +79,7 @@ successes =
     ("2 + 2", "const main = 2 + 2;\n"),
     ("10 - 2", "const main = 10 - 2;\n"),
     ("\"dog\" <> \"log\"", "const main = \"dog\" + \"log\";\n"),
-    ("{ fn: (\\a -> let d = 1 in a) }", "const main = { fn: a => { const d = 1;\nreturn a } }")
+    ("{ fn: (\\a -> let d = 1 in a) }", "const main = { fn: a => { const d = 1;\nreturn a } };\n")
   ]
 
 testIt :: (Text, Javascript) -> Spec
