+++ test/Test/Parser/Syntax.hs
@@ -9,6 +9,7 @@ import Data.Either (isLeft, isRight)
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map as M
 import Data.Text (Text)
+import Language.Mimsa.ExprUtils
 import Language.Mimsa.Parser
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Identifiers
